backend/requirements.txt.txt
fastapi
uvicorn
httpx
pydantic
python-multipart
pdfplumber
reportlab
jinja2
duckduckgo_search
python-docx